Fixed the Schedular to prepare by Date for One week
Fixed the Schedular to prepare by Date for One week

+++ src/main/java/com/gaksvytech/fieldservice/entity/Events.java
@@ -39,7 +39,7 @@ public class Events {
 	@GeneratedValue
 	private long id;
 
-	private long zoneId;
+	private int zoneId;
 
 	private String name;
